requirements.txt

charset-normalizer==2.1.0
Cython==0.29.30
filelock==3.7.1
gdown==4.5.1
h5py==3.7.0
hdf5storage==0.1.18
idna==3.3
nms==0.1.6
numpy==1.23.0
opencv-python==4.6.0.66
Pillow==9.2.0
PySocks==1.7.1
requests==2.28.1
scipy==1.8.1
six==1.16.0
soupsieve==2.3.2.post1
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.9
utils==1.0.1